Modification to image gradient to filter to compute gradients wrt scanner coords.
Changed default smoothing stdev to 1 voxel

+++ lib/image/adapter/gradient1D.h
@@ -44,9 +44,9 @@ namespace MR
             axis_ = axis;
           }
 
-          ssize_t value () {
+          float value () {
             const ssize_t pos = (*this)[axis_];
-            ssize_t val = 0.0;
+            float val = 0.0;
 
             if (pos == 0) {
               val = parent_vox.value();
